début hitbox

+++ LeJeuShared/src/main/java/fr/devops/shared/ingame/control/IPlayerController.java
@@ -4,7 +4,7 @@ import fr.devops.shared.service.IService;
 
 public interface IPlayerController extends IService{
 
-	public byte getXAxis();
+	public int getXAxis();
 	
 	public boolean isJumping();
